requirements.txt
alembic==1.4.2
astroid==2.4.2
autopep8==1.5.3
bcrypt==3.1.7
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dnspython==1.16.0
email-validator==1.1.1
flake8==3.8.3
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.9
importlib-metadata==1.6.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.1
Jinja2==2.11.2
joblib==0.15.1
langdetect==1.0.8
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
newsapi-python==0.2.6
nltk==3.5
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.5
parso==0.7.0
Pillow==7.1.2
pluggy==0.13.1
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.0.2
pyflakes==2.2.0
pylint==2.5.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-jsonrpc-server==0.3.4
python-language-server==0.33.2
pytz==2020.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.17.0
scikit-learn==0.23.1
scipy==1.5.0
six==1.15.0
sklearn==0.0
snowballstemmer==2.0.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.46.1
tweepy==3.8.0
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
yapf==0.30.0
zipp==3.1.0
